requirements.txt
anyio==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
dnspython==2.2.0
Flask==2.0.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.23.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==9.0.1
pymongo==4.0.1
python-dotenv==0.19.2
git+https://github.com/ifahadone/pytube
qrcode==7.3.1
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
urllib3==1.26.8
Werkzeug==2.0.2
Zenora==0.0.3
